Move configuration to variable, start support for topojson
Move userVariables to be a string variable.  This way the customer can set configuration without modifying code, will make it easier for future upgrades.

Started detection and support for topoJSON, still in progress.  Using the omnivore plugin from mapbox,
this vis is different than the original US States in ZD.  The old viz only converted topojson features on demand,
while we can't do that since there is no guarantee that the construction will be consistent between layers.

Also added a max extent/max zoom to keep the user pinned to the area of interest.