Revise filtering stratgey based on fixed colors variable
If the fixed colors variable is true (default) then the filters are only recalculated when the layer is changed on zoom.  The filter is set to all values in the GeoJSON for that layer, so the min/max are not recalculated when the user pans the map (or zooms w/out triggering a layer change).  Zoomdata is not re-queried on pan, to the color values are not recalculated and stay consistent.